Bug 1663940: Ensure the unwriteable margins are populated in the print settings. r=emilio
This updates the settings with the correct paper ID as well.
It also removes the printable size fields from the Windows settings now that
unwriteable margins are being used consistently.

Differential Revision: https://phabricator.services.mozilla.com/D89911

+++ widget/nsPrinterBase.cpp
@@ -150,7 +150,7 @@ NS_IMETHODIMP nsPrinterBase::GetPaperList(JSContext* aCx,
                                      &nsPrinterBase::PaperList);
 }
 
-void nsPrinterBase::QueryMarginsForPaper(Promise& aPromise, uint64_t aPaperId) {
+void nsPrinterBase::QueryMarginsForPaper(Promise& aPromise, short aPaperId) {
   return SpawnPrintBackgroundTask(*this, aPromise, "MarginsForPaper"_ns,
                                   &nsPrinterBase::GetMarginsForPaper, aPaperId);
 }
