+++ widget/nsPrinterBase.h
@@ -45,7 +45,7 @@ class nsPrinterBase : public nsIPrinter {
   nsPrinterBase(const nsPrinterBase&) = delete;
   nsPrinterBase(nsPrinterBase&&) = delete;
 
-  void QueryMarginsForPaper(Promise&, uint64_t aPaperId);
+  void QueryMarginsForPaper(Promise&, short aPaperId);
 
   /**
    * Caches the argument by copying it into mPrintSettingsInitializer.
@@ -95,7 +95,7 @@ class nsPrinterBase : public nsIPrinter {
   virtual bool SupportsMonochrome() const = 0;
   virtual bool SupportsCollation() const = 0;
   virtual nsTArray<mozilla::PaperInfo> PaperList() const = 0;
-  virtual MarginDouble GetMarginsForPaper(uint64_t aPaperId) const = 0;
+  virtual MarginDouble GetMarginsForPaper(short aPaperId) const = 0;
 
  private:
   mozilla::EnumeratedArray<AsyncAttribute, AsyncAttribute::Last,
