+++ widget/nsPrinterCUPS.h
@@ -26,7 +26,7 @@ class nsPrinterCUPS final : public nsPrinterBase {
   bool SupportsMonochrome() const final;
   bool SupportsCollation() const final;
   nsTArray<mozilla::PaperInfo> PaperList() const final;
-  MarginDouble GetMarginsForPaper(uint64_t) const final {
+  MarginDouble GetMarginsForPaper(short) const final {
     MOZ_ASSERT_UNREACHABLE(
         "The CUPS API requires us to always get the margin when fetching the "
         "paper list so there should be no need to query it separately");
